solve todos, fetch all db entries, make page size configurable

+++ server/model/pagination.go
@@ -1,6 +1,6 @@
 package model
 
 type PaginationData struct {
-	Page    int64
-	PerPage int64
+	Page    int
+	PerPage int
 }
